server/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
appdirs==1.4.4
attrs==19.3.0
boto3==1.14.22
botocore==1.17.22
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
coverage==5.2
databricks-cli==0.11.0
dill==0.3.2
distlib==0.3.1
docker==4.2.2
docutils==0.15.2
dotmap==1.3.0
elastic-apm==5.8.1
elasticsearch==7.8.0
entrypoints==0.3
farm==0.4.6
farm-haystack==0.3.0
fastapi==0.59.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
future==0.18.2
gitdb==4.0.5
GitPython==3.1.7
gunicorn==20.0.4
h11==0.9.0
h5py==2.10.0
httptools==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
Keras==2.4.3
langdetect==1.0.8
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
mlflow==1.0.0
numpy==1.19.0
packaging==20.4
pandas==1.0.5
pluggy==0.13.1
protobuf==3.12.2
psutil==5.7.2
psycopg2-binary==2.8.5
py==1.9.0
pydantic==1.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-docx==0.8.10
python-editor==1.0.4
python-multipart==0.0.5
pytz==2020.1
PyYAML==5.3.1
querystring-parser==1.2.4
regex==2020.7.14
requests==2.24.0
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.5.1
sentencepiece==0.1.91
seqeval==0.0.12
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smmap==3.0.4
SQLAlchemy==1.3.18
sqlparse==0.3.1
starlette==0.13.4
tabulate==0.8.7
threadpoolctl==2.1.0
tokenizers==0.8.1rc1
toml==0.10.1
torch==1.5.1
tox==3.17.1
tqdm==4.48.0
transformers==3.0.2
urllib3==1.25.9
uvicorn==0.11.6
uvloop==0.14.0
virtualenv==20.0.27
websocket-client==0.57.0
websockets==8.1
Werkzeug==0.16.1
wget==3.2
